# docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Tether Local Proxy API",
    "version": "0.1.0",
    "description": "Local REST API exposed by the Tether/Tether desktop proxy process."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8080",
      "description": "Default local proxy"
    }
  ],
  "paths": {
...
    "/api/traces/current": {
      "get": {
        "summary": "Read the current trace graph",
        "tags": ["traces"],
        "parameters": [
          {
            "name": "session_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "UI-readable trace snapshot",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TraceSnapshot"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Clear all traces and reset the live session",
        "tags": ["traces"],
        "responses": {
          "204": {
            "description": "Trace storage cleared"
          }
        }
      }
    },
    "/api/traces/{id}/output": {
      "patch": {
        "summary": "Edit a trace node output and mark downstream nodes stale",
        "tags": ["traces"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditOutputRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Invalidated descendants",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidationResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/traces/{id}/downstream": {
      "get": {
        "summary": "List downstream trace nodes for a node",
        "tags": ["traces"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Downstream node ids",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DownstreamResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/traces/{id}/replay": {
      "post": {
        "summary": "Replay a retained trace node request",
        "tags": ["traces"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Replay result and re-invalidated descendants",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReplayResult"
                }
              }
            }
          },
          "409": {
            "description": "Node is not replayable"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "TraceSession": {
        "type": "object",
        "required": ["id", "title", "trigger", "started_at"],
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "trigger": {
            "type": "string"
          },
          "started_at": {
            "type": "string"
          }
        }
      },
...
      "TraceSnapshot": {
        "type": "object",
        "required": ["nodes", "stale_node_ids"],
        "properties": {
          "session": {
            "$ref": "#/components/schemas/TraceSession",
            "nullable": true
          },
          "nodes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AgentNode"
            }
          },
          "stale_node_ids": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "EditOutputRequest": {
        "type": "object",
        "required": ["output"],
        "properties": {
          "output": {
            "type": "string"
          }
        }
      },
      "InvalidationResult": {
        "type": "object",
        "required": ["node_id", "reason", "previous_output_hash", "output_hash", "invalidated"],
        "properties": {
          "node_id": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          },
          "previous_output_hash": {
            "type": "string"
          },
          "output_hash": {
            "type": "string"
          },
          "invalidated": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "DownstreamResult": {
        "type": "object",
        "required": ["node_id", "downstream"],
        "properties": {
          "node_id": {
            "type": "string"
          },
          "downstream": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "ReplayResult": {
        "type": "object",
        "required": [
          "node_id",
          "reason",
          "previous_output_hash",
          "output_hash",
          "status_code",
          "cost",
          "tokens_in",
          "tokens_out",
          "invalidated"
        ],
        "properties": {
          "node_id": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          },
          "previous_output_hash": {
            "type": "string"
          },
          "output_hash": {
            "type": "string"
          },
          "status_code": {
            "type": "integer",
            "format": "int32"
          },
          "cost": {
            "type": "string"
          },
          "tokens_in": {
            "type": "integer",
            "format": "int64"
          },
          "tokens_out": {
            "type": "integer",
            "format": "int64"
          },
          "invalidated": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "AgentNode": {
        "type": "object",
        "required": [
          "id",
          "trace_id",
          "tool_use_ids",
          "context_inputs",
          "input_hash",
          "stale",
          "agent_name",
          "depth",
          "step_name",
          "timestamp",
          "provider",
          "model",
          "cost",
          "latency",
          "latency_ms",
          "bar_percent",
          "tokens_in",
          "tokens_out",
          "request_id",
          "cache_status",
          "status",
          "prompt",
          "response",
          "output_hash"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "trace_id": {
            "type": "string"
          },
          "parent_span_id": {
            "type": "string",
            "nullable": true
          },
          "tool_use_ids": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "context_inputs": {
            "$ref": "#/components/schemas/ContextInputs"
          },
          "input_hash": {
            "type": "string"
          },
          "stale": {
            "type": "boolean"
          },
          "agent_name": {
            "type": "string"
          },
          "depth": {
            "type": "integer",
            "format": "int64"
          },
          "step_name": {
            "type": "string"
          },
          "timestamp": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "cost": {
            "type": "string"
          },
          "latency": {
            "type": "string"
          },
          "latency_ms": {
            "type": "integer",
            "format": "int64"
          },
          "bar_percent": {
            "type": "number",
            "format": "double"
          },
          "tokens_in": {
            "type": "integer",
            "format": "int64"
          },
          "tokens_out": {
            "type": "integer",
            "format": "int64"
          },
          "request_id": {
            "type": "string"
          },
          "cache_status": {
            "type": "string",
            "enum": ["hit", "miss", "codex-log"]
          },
          "temperature": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": ["success", "cached", "running", "error"]
          },
          "prompt": {
            "$ref": "#/components/schemas/AgentPrompt"
          },
          "response": {
            "$ref": "#/components/schemas/AgentResponse"
          },
          "output_hash": {
            "type": "string"
          },
          "error": {
            "$ref": "#/components/schemas/AgentError",
            "nullable": true
          }
        }
      },
      "ContextInputs": {
        "type": "object",
        "required": ["sources", "withheld", "input_hash"],
        "properties": {
          "sources": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ContextSource"
            }
          },
          "withheld": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "input_hash": {
            "type": "string"
          }
        }
      },
      "ContextSource": {
        "type": "object",
        "required": ["kind", "path_or_id", "hash", "size_bytes"],
        "properties": {
          "kind": {
            "type": "string"
          },
          "path_or_id": {
            "type": "string"
          },
          "hash": {
            "type": "string"
          },
          "size_bytes": {
            "type": "integer",
            "format": "int64"
          },
          "body": {
            "type": "string"
          }
        }
      },
      "AgentPrompt": {
        "type": "object",
        "required": ["system", "user"],
        "properties": {
          "system": {
            "type": "string"
          },
          "user": {
            "type": "string"
          }
        }
      },
      "AgentResponse": {
        "type": "object",
        "required": ["language", "text"],
        "properties": {
          "language": {
            "type": "string"
          },
          "text": {
            "type": "string"
          }
        }
      },
      "AgentError": {
        "type": "object",
        "required": ["code", "message", "detail"],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "detail": {
            "type": "string"
          }
        }
      },
...
    }
  }
}